target/hexagon: use different HVX semantics among revs
When importing v79 HVX, we ended up breaking the simulation for v75
binaries in QEMU as the semantics changed in some of the instructions,
specially for QFloat, making the output values from v75 and v79 differ.
So QEMU can no longer simulate v75 programs correctly. Let's fix that by
re-introducing the v75 semantics and letting QEMU use the right version
depending on the running revision. For now, the instruction descriptions
and attributes are always set to the latest revision, for convenience.

Signed-off-by: Matheus Tavares Bernardino <quic_mathbern@quicinc.com>
Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>