target/hexagon: Refactor the k0,tlb lock
This design of the locks is recommended by community and will be easier to
upstream.  It's much simpler to reason about behavior when vCPUs don't modify
one another's state when unlocking.

Fairness behavior of the locks as specified by the system architecture
should be preserved, due to the fair host schedulers of the various host OSs,
when the thundering herd races to get the lock.

Link: https://lore.kernel.org/qemu-devel/7fc5e75b-4968-4df8-818c-ab2c4c758531@linaro.org/
Suggested-by: Richard Henderson <richard.henderson@linaro.org>
Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>